Rename component to be singular
